+++ pootle/apps/pootle_store/views.py
@@ -8,6 +8,7 @@
 # AUTHORS file for copyright and authorship information.
 
 import copy
+import time
 
 from translate.lang import data
 
@@ -30,7 +31,6 @@ from pootle.core.delegate import search_backend
 from pootle.core.exceptions import Http400
 from pootle.core.http import JsonResponse, JsonResponseBadRequest
 from pootle.core.mail import send_mail
-from pootle.core.utils import dateformat
 from pootle.core.views import (BaseBrowseDataJSON, BasePathDispatcherView,
                                PootleJSON)
 from pootle.i18n.gettext import ugettext as _
@@ -358,15 +358,12 @@ class UnitTimelineJSON(PootleUnitJSON):
     def get_entries_group_data(self, context):
         result = []
         for entry_group in context['entries_group']:
-            display_dt = entry_group['datetime']
-            if display_dt is not None:
-                display_dt = dateformat.format(display_dt)
-                iso_dt = entry_group['datetime'].isoformat()
-            else:
-                iso_dt = None
+            timestamp = None
+            entry_datetime = entry_group['datetime']
+            if entry_datetime is not None:
+                timestamp = time.mktime(entry_datetime.timetuple())
             result.append({
-                "display_datetime": display_dt,
-                "iso_datetime": iso_dt,
+                'timestamp': timestamp,
             })
         return result